pyproject.toml
[project]
name = "bitrouter-bench"
version = "0.1.0"
description = "Benchmark harness for evaluating BitRouter routing quality vs direct API calls"
requires-python = ">=3.11"
license = "Apache-2.0"
dependencies = [
    "httpx>=0.27",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "click>=8.0",
    "python-frontmatter>=1.1",
    "tabulate>=0.9",
    "rich>=13.0",
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "sse-starlette>=2.0",
]

[project.scripts]
bench = "bitrouter_bench.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bitrouter_bench"]
